WIP set outputs again
