+++ conda_forge_tick/migrators/migration_yaml.py
@@ -14,8 +14,9 @@ import ruamel.yaml as yaml
 from conda_forge_tick.contexts import FeedstockContext
 from conda_forge_tick.migrators.core import GraphMigrator, MiniMigrator, Migrator
 from conda_forge_tick.xonsh_utils import indir
-from conda_forge_tick.utils import eval_cmd
-from ..utils import pluck, pin_sep_pat
+from conda_forge_tick.utils import eval_cmd, pluck
+from conda_forge_tick.make_graph import get_deps_from_outputs_lut
+from conda_forge_tick.feedstock_parser import PIN_SEP_PAT
 
 if typing.TYPE_CHECKING:
     from ..migrators_types import (
@@ -422,7 +423,7 @@ class MigrationYamlCreator(Migrator):
 
 
 def _req_is_python(req):
-    return pin_sep_pat.split(req)[0].strip().lower() == "python"
+    return PIN_SEP_PAT.split(req)[0].strip().lower() == "python"
 
 
 def _all_noarch(attrs, only_python=False):
@@ -488,13 +489,11 @@ def create_rebuild_graph(
         )
         # get host/build, run and test and launder them through outputs
         # this should fix outputs related issues (eg gdal)
-        rq = set(
-            map(
-                lambda x: gx.graph["outputs_lut"].get(x, x),
-                (host or build)
-                | requirements.get("run", set())
-                | requirements.get("test", set()),
-            ),
+        rq = get_deps_from_outputs_lut(
+            (host or build)
+            | requirements.get("run", set())
+            | requirements.get("test", set()),
+            gx.graph["outputs_lut"],
         )
 
         for e in list(total_graph.in_edges(node)):
